+++ doc/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 ======
 
+v0.14.1 - 04.11.2023
+----
+* Fixes in the scene changing mechanism;
+* New features for building scenes: FixText and BaseButton;
+* Added Title Scene (currently hidden).
+
 v0.14.0 - 02.11.2023
 ----
 * Preparation of the scene changing mechanism.